+++ src/lmdesktopplus/__init__.py
@@ -1,3 +1,3 @@
 """LMDesktopPlus machine UI."""
 
-__version__ = "0.6.4"
+__version__ = "0.6.5"